(PC-13180)[API] refactor: move shared users routes to pro
