+++ api/src/pcapi/routes/shared/__init__.py
@@ -8,4 +8,3 @@ def install_routes(app: Flask) -> None:
     # pylint: disable=unused-import
     from . import features
     from . import passwords
-    from . import users
